Dretax

+++ Garhal/events.c
@@ -9,18 +9,18 @@
 // Find our required PE image.
 PLOAD_IMAGE_NOTIFY_ROUTINE ImageLoadCallback(PUNICODE_STRING FullImageName, HANDLE ProcessId, PIMAGE_INFO ImageInfo)
 {
-	//DbgPrintEx(0, 0, "We received a load from: %ls \n", FullImageName->Buffer);
+	//DebugMessageNormal("We received a load from: %ls \n", FullImageName->Buffer);
 
 	// Compare our string to input
 	if (wcsstr(FullImageName->Buffer, L"\\csgo\\bin\\client.dll"))
 	{
 		DebugMessageNormal("CSGO client.dll found!\n");
 		DebugMessageNormal("Loaded Name: %ls \n", FullImageName->Buffer);
-		DebugMessageNormal("Loaded To Process: %d \n", ProcessId);
+		DebugMessageNormal("Loaded To Process: %d \n", (ULONG) ProcessId);
 
-		ClientAddress = ImageInfo->ImageBase;
-		ClientSize = ImageInfo->ImageSize;
-		csgoId = ProcessId;
+		ClientAddress = (ULONG) ImageInfo->ImageBase;
+		ClientSize = (ULONG) ImageInfo->ImageSize;
+		csgoId = (ULONG) ProcessId;
 
 		// Free, and re-init the vector every time CSGO loads up.
 		vector_free(&CSRSSList);
@@ -34,8 +34,8 @@ PLOAD_IMAGE_NOTIFY_ROUTINE ImageLoadCallback(PUNICODE_STRING FullImageName, HAND
 	{
 		DebugMessageNormal("CSGO Engine.dll found!\n");
 
-		EngineAddress = ImageInfo->ImageBase;
-		EngineSize = ImageInfo->ImageSize;
+		EngineAddress = (ULONG) ImageInfo->ImageBase;
+		EngineSize = (ULONG) ImageInfo->ImageSize;
 	}
 
 	return STATUS_SUCCESS;
@@ -51,7 +51,7 @@ PCREATE_PROCESS_NOTIFY_ROUTINE_EX ProcessNotifyCallbackEx(HANDLE parentId, HANDL
 		if (wcsstr(notifyInfo->ImageFileName->Buffer, L"\\GarhalController.exe"))
 		{
 			DebugMessageNormal("Bomb has been planted!\n");
-			ControllerID = processId;
+			ControllerID = (ULONG) processId;
 			DebugMessageNormal("Controller ProcessID: %d\r\n", ControllerID);
 
 			if (EnableProcessHiding == 1)
@@ -64,7 +64,7 @@ PCREATE_PROCESS_NOTIFY_ROUTINE_EX ProcessNotifyCallbackEx(HANDLE parentId, HANDL
 		}
 		else if (wcsstr(notifyInfo->ImageFileName->Buffer, L"\\GarhalRankDisplayer.exe"))
 		{
-			RankReaderID = processId;
+			RankReaderID = (ULONG) processId;
 			DebugMessageNormal("RankReaderID ProcessID: %d\r\n", RankReaderID);
 
 			if (EnableProcessHiding == 1)
@@ -78,7 +78,7 @@ PCREATE_PROCESS_NOTIFY_ROUTINE_EX ProcessNotifyCallbackEx(HANDLE parentId, HANDL
 	}
 	else
 	{
-		ULONG ProcID = (ULONG)processId;
+		ULONG ProcID = (ULONG) processId;
 		if (ControllerID == ProcID)
 		{
 			DebugMessageNormal("Controller Shutdown detected, disabling protection. %d\r\n", ControllerID);
